feat(dev-lang/spidermonkey): add package & patches
This package was added so python 3.12 could be enabled. Upstream still
doesn't fully support python 3.12, but they're close enough that we can
backport the necessary patches to get it working.

The spidermonkey-115.3.1-Python-3.12-support.patch is a combination of
the several upstream commits (which didn't all apply cleanly) and can be
found in the mozilla/gecko-dev GitHub repo.

The commits are:
- 7293cfa
- bea8453
- 272077c
- 65ae752
- 766f65d
- 7cc44b5
- ca039c7

A number of tests started failing after this patch was applied, but the
built package still seems to work fine when used to test dev-libs/gjs.
So they were added to the list of expected failures.

Refs: #1
Signed-off-by: Jaremy Hatler <hatler.jaremy@gmail.com>